+++ frontend/src/routes/_authenticated/folders/$folderName/template-policies/$policyName.tsx
@@ -133,8 +133,8 @@ export function FolderTemplatePolicyDetailPage({
                 {orgName}
               </Link>
               {' / '}
-              <Link to="/organizations/$orgName/resources" params={{ orgName }} className="hover:underline">
-                Resources
+              <Link to="/organizations/$orgName/projects" params={{ orgName }} className="hover:underline">
+                Projects
               </Link>
               {' / '}
               <Link