+++ frontend/src/routes/_authenticated/folders/$folderName/template-policies/new.tsx
@@ -61,8 +61,8 @@ export function CreateFolderTemplatePolicyPage({
               {orgName}
             </Link>
             {' / '}
-            <Link to="/organizations/$orgName/resources" params={{ orgName }} className="hover:underline">
-              Resources
+            <Link to="/organizations/$orgName/projects" params={{ orgName }} className="hover:underline">
+              Projects
             </Link>
             {' / '}
             <Link