+++ frontend/src/routes/_authenticated/folders/$folderName/template-policy-bindings/$bindingName.tsx
@@ -127,11 +127,11 @@ export function FolderTemplatePolicyBindingDetailPage({
               </Link>
               {' / '}
               <Link
-                to="/organizations/$orgName/resources"
+                to="/organizations/$orgName/projects"
                 params={{ orgName }}
                 className="hover:underline"
               >
-                Resources
+                Projects
               </Link>
               {' / '}
               <Link