+++ frontend/src/routes/_authenticated/folders/$folderName/templates/new.tsx
@@ -43,8 +43,8 @@ export function CreateFolderTemplatePage({ folderName: propFolderName }: { folde
               {orgName}
             </Link>
             {' / '}
-            <Link to="/organizations/$orgName/resources" params={{ orgName }} className="hover:underline">
-              Resources
+            <Link to="/organizations/$orgName/projects" params={{ orgName }} className="hover:underline">
+              Projects
             </Link>
             {' / '}
             <Link